+++ .vscode/settings.json
@@ -16,6 +16,7 @@
     "yarn.lock": true
   },
   "eslint.nodePath": ".yarn/sdks",
+  "eslint.packageManager": "yarn",
   "prettier.prettierPath": ".yarn/sdks/prettier/index.js",
   "typescript.tsdk": ".yarn/sdks/typescript/lib",
   "typescript.enablePromptUseWorkspaceTsdk": true,
