+++ package.json
@@ -25,41 +25,41 @@
   },
   "dependencies": {
     "@apollo/client": "^3.6.9",
-    "date-fns": "^2.29.1",
-    "dotenv": "^16.0.1",
-    "graphql": "^16.5.0",
+    "date-fns": "^2.29.2",
+    "dotenv": "^16.0.2",
+    "graphql": "^16.6.0",
     "html5-qrcode": "^2.2.1",
-    "next": "12.2.3",
-    "next-pwa": "^5.5.5",
+    "next": "12.2.5",
+    "next-pwa": "^5.6.0",
     "normalize.css": "^8.0.1",
     "qrcode": "^1.5.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.34.0",
+    "react-hook-form": "^7.34.2",
     "react-icons": "^4.4.0",
     "react-timer-hook": "^3.0.5",
-    "react-toastify": "^9.0.7",
-    "recoil": "^0.7.4",
+    "react-toastify": "^9.0.8",
+    "recoil": "^0.7.5",
     "sharp": "^0.30.7",
     "styled-components": "^5.3.5",
     "swr": "^1.3.0",
     "ts-node": "^10.9.1"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^2.11.3",
-    "@graphql-codegen/typescript": "^2.7.2",
-    "@graphql-codegen/typescript-apollo-client-helpers": "^2.2.2",
-    "@graphql-codegen/typescript-operations": "^2.5.2",
-    "@graphql-codegen/typescript-react-apollo": "^3.3.2",
+    "@graphql-codegen/cli": "^2.12.0",
+    "@graphql-codegen/typescript": "^2.7.3",
+    "@graphql-codegen/typescript-apollo-client-helpers": "^2.2.3",
+    "@graphql-codegen/typescript-operations": "^2.5.3",
+    "@graphql-codegen/typescript-react-apollo": "^3.3.3",
     "@svgr/webpack": "^6.3.1",
     "@testing-library/react": "^13.3.0",
     "@types/gtag.js": "^0.0.10",
-    "@types/jest": "^28.1.6",
-    "@types/node": "^18.6.3",
-    "@types/qrcode": "^1.4.2",
-    "@types/react": "^18.0.15",
+    "@types/jest": "^28.1.8",
+    "@types/node": "^18.7.14",
+    "@types/qrcode": "^1.5.0",
+    "@types/react": "^18.0.18",
     "@types/react-dom": "^18.0.6",
-    "@types/styled-components": "^5.1.25",
+    "@types/styled-components": "^5.1.26",
     "eslint": "^8.21.0",
     "eslint-config-next": "12.2.3",
     "eslint-config-prettier": "^8.5.0",
