+++ src/pages/_offline.tsx
@@ -1,0 +1,9 @@
+import PageHead from 'src/components/PageHead'
+
+export default function OfflinePage() {
+  return (
+    <PageHead title=" - 자유담" description="">
+      PWA 오프라인. 인터넷에 연결해주세요
+    </PageHead>
+  )
+}
