+++ src/pages/post/[id]/index.tsx
@@ -2,7 +2,6 @@ import LoginLink from 'src/components/atoms/LoginLink'
 import Image from 'next/future/image'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
-import { MouseEvent, ReactNode } from 'react'
 import { toast } from 'react-toastify'
 import { useRecoilValue } from 'recoil'
 import { toastApolloError } from 'src/apollo/error'
@@ -15,18 +14,17 @@ import {
   useToggleLikingPostMutation,
 } from 'src/graphql/generated/types-and-hooks'
 import Navigation from 'src/layouts/Navigation'
-import { Skeleton, flexBetween, flexCenter } from 'src/styles'
+import { flexBetween, flexCenter } from 'src/styles'
 import { theme } from 'src/styles/global'
 import BackArrowIcon from 'src/svgs/back-arrow.svg'
 import CommentIcon from 'src/svgs/CommentIcon'
 import HeartIcon from 'src/svgs/HeartIcon'
-import ShareIcon from 'src/svgs/ShareIcon'
 import ThreeDotsIcon from 'src/svgs/three-dots.svg'
-import { stopPropagation } from 'src/utils'
 import { currentUser } from 'src/utils/recoil'
 import styled from 'styled-components'
 
 import { borderRadiusCircle } from '..'
+import CommentCard from 'src/components/CommentCard'
 
 const description = ''
 
@@ -187,182 +185,6 @@ export default function PostPage() {
   )
 }
 
-type Props2 = {
-  comment: Post
-  showSharedPost?: boolean
-}
-
-export function CommentCard({ comment, showSharedPost }: Props2) {
-  const comments = comment.comments
-
-  return (
-    <Card>
-      <CommentContent comment={comment} showParentAuthor showSharedPost={showSharedPost}>
-        {comments && <VerticalLine />}
-        {comments?.map((comment, i) => (
-          <CommentContent key={comment.id} comment={comment}>
-            {[comments?.length - 1 !== i && <VerticalLine key={i} />]}
-          </CommentContent>
-        ))}
-      </CommentContent>
-    </Card>
-  )
-}
-
-export function PostLoadingCard() {
-  return (
-    <Card>
-      <FlexColumn>
-        <Skeleton width="40px" height="40px" borderRadius="50%" />
-      </FlexColumn>
-      <GridSmallGap>
-        <Skeleton width="40%" />
-        <Skeleton />
-        <Skeleton width="80%" />
-        <GridColumn4>
-          <Width>
-            <HeartIcon /> <Skeleton width="1rem" />
-          </Width>
-          <Width>
-            <CommentIcon /> <Skeleton width="1rem" />
-          </Width>
-          <Width>
-            <ShareIcon /> <Skeleton width="1rem" />
-          </Width>
-          <div>기타</div>
-        </GridColumn4>
-      </GridSmallGap>
-    </Card>
-  )
-}
-
-type Props = {
-  children?: ReactNode[]
-  comment: Post
-  showParentAuthor?: boolean
-  showSharedPost?: boolean
-}
-
-function CommentContent({ children, comment, showParentAuthor, showSharedPost }: Props) {
-  const author = comment.author
-  const parentAuthor = comment.parentAuthor
-  const sharedPost = comment.sharingPost
-
-  // 해당 글로 이동하기
-  const router = useRouter()
-
-  function goToPostPage() {
-    router.push(`/post/${comment.id}`)
-  }
-
-  function goToUserPage(e: MouseEvent<HTMLElement>) {
-    e.stopPropagation()
-
-    if (author) {
-      router.push(`/@${author.name}`)
-    }
-  }
-
-  // 좋아요
-  const [toggleLikingPostMutation, { loading }] = useToggleLikingPostMutation({
-    onError: toastApolloError,
-    variables: { id: comment.id },
-  })
-
-  function toggleLikingPost(e: any) {
-    e.preventDefault()
-    e.stopPropagation()
-    toggleLikingPostMutation()
-  }
-
-  // 댓글 달기
-  function showPostCreationModal(e: any) {
-    e.preventDefault()
-    e.stopPropagation()
-  }
-
-  return (
-    <>
-      <FlexColumn onClick={goToPostPage}>
-        <Image
-          src={author?.imageUrl ?? '/images/shortcut-icon.webp'}
-          alt="profile"
-          width="40"
-          height="40"
-          onClick={goToUserPage}
-          style={borderRadiusCircle}
-        />
-        {children?.[0]}
-      </FlexColumn>
-      <GridSmallGap onClick={goToPostPage}>
-        <FlexBetween>
-          <div>
-            <Bold disabled={!author} onClick={goToUserPage}>
-              {author?.nickname ?? '탈퇴한 사용자'}
-            </Bold>{' '}
-            {author && (
-              <LineLink href={`/@${author.name}`} onClick={stopPropagation}>
-                <GreyInlineH5>@{author.name}</GreyInlineH5>
-              </LineLink>
-            )}
-            {' · '}
-            <span>{new Date(comment.creationTime).toLocaleDateString()}</span>
-            <span>{comment.updateTime && '(수정됨)'}</span>
-          </div>
-          <ThreeDotsIcon width="1rem" />
-        </FlexBetween>
-
-        {showParentAuthor && parentAuthor && author && parentAuthor.name !== author.name && (
-          <GreyInlineH5>
-            Replying to{' '}
-            <Link href={`/@${parentAuthor.name}`} onClick={stopPropagation}>
-              @{parentAuthor.name}
-            </Link>
-          </GreyInlineH5>
-        )}
-
-        <p>
-          {comment.deletionTime
-            ? `${new Date(comment.deletionTime).toLocaleString()} 에 삭제된 글이에요`
-            : comment.content}
-        </p>
-
-        {showSharedPost && sharedPost && <SharedPostCard sharedPost={sharedPost as Post} />}
-
-        <GridColumn4>
-          <div>
-            <Button color={theme.error} onClick={toggleLikingPost} selected={comment.isLiked}>
-              <HeartIcon /> <span>{comment.likeCount}</span>
-            </Button>
-          </div>
-          <div>
-            <Button
-              color={theme.primaryText}
-              onClick={showPostCreationModal}
-              selected={comment.doIComment}
-            >
-              <CommentIcon /> <span>{comment.commentCount}</span>
-            </Button>
-          </div>
-          <SharingPostButton post={comment} sharedPost={comment} />
-          <div>기타</div>
-        </GridColumn4>
-      </GridSmallGap>
-
-      {children?.[1]}
-    </>
-  )
-}
-
-export const Card = styled.li`
-  display: grid;
-  grid-template-columns: auto 1fr;
-  gap: 0.5rem;
-
-  border: 1px solid ${(p) => p.theme.shadow};
-  padding: 0.8rem 1rem;
-`
-
 const Sticky = styled.div`
   position: sticky;
   top: 0;
@@ -382,12 +204,6 @@ const Grid = styled.div`
   display: grid;
   gap: 1rem;
   padding: 1rem;
-`
-
-export const GridSmallGap = styled.div`
-  display: grid;
-  gap: 0.5rem;
-  cursor: pointer;
 `
 
 const FlexBetween = styled.div`
@@ -418,11 +234,11 @@ export const GreyH5 = styled.h5`
   font-weight: 400;
 `
 
-const GreyInlineH5 = styled(GreyH5)`
+export const GreyInlineH5 = styled(GreyH5)`
   display: inline-block;
 `
 
-const GridColumn4 = styled.div`
+export const GridColumn4 = styled.div`
   display: grid;
   grid-template-columns: 1fr 1fr 1fr 1fr;
   gap: 1rem;
@@ -439,19 +255,7 @@ const GridColumn4Center = styled(GridColumn4)`
   }
 `
 
-const FlexColumn = styled.div`
-  display: flex;
-  flex-flow: column;
-  gap: 0.5rem;
-`
-
-const VerticalLine = styled.div`
-  border-left: 1px solid #888;
-  margin: auto;
-  flex-grow: 1;
-`
-
-const Bold = styled.b<{ disabled: boolean }>`
+export const Bold = styled.b<{ disabled: boolean }>`
   color: ${(p) => (p.disabled ? p.theme.primaryAchromatic : '#000')};
   cursor: ${(p) => (p.disabled ? 'not-allowed' : 'pointer')};
 `
@@ -493,11 +297,5 @@ export const Button = styled.button<{ color: string; selected: boolean }>`
     > span {
       color: ${(p) => p.color};
     }
-  }
-`
-
-const Width = styled.div`
-  > svg {
-    width: 1rem;
   }
 `
